+++ gitlab/tests/test_gitlab_http_methods.py
@@ -1,4 +1,5 @@
 import pytest
+import requests
 
 from httmock import HTTMock, urlmatch, response
